Moved TableView to its own file
