+++ FWCore/SOA/interface/Table.h
@@ -34,9 +34,6 @@
 namespace edm {
 namespace soa {
   
-  template<typename... Args> class TableView;
-  
-
   template <typename... Args>
   class Table {
   public:
@@ -106,9 +103,6 @@ namespace soa {
     MutableColumnValues<typename U::type> column() {
       return MutableColumnValues<typename U::type>{static_cast<typename U::type*>(columnAddress<U>()), m_size};
     }
-    
-    template<typename... U>
-    TableView<U...> view() const;
     
     const_iterator begin() const { return const_iterator{m_values}; }
     const_iterator end() const { return const_iterator{m_values,size()}; }
@@ -284,72 +278,7 @@ namespace soa {
     
   };
   
-  
-  template <typename... Args>
-  class TableView {
     
-  public:
-    using Layout = std::tuple<Args...>;
-    static constexpr const size_t kNColumns = sizeof...(Args);
-    using const_iterator = ConstTableItr<Args...>;
-    
-    template <typename... OArgs>
-    TableView( Table<OArgs...> const& iTable):
-    m_size(iTable.size()) {
-      fillArray<0>(iTable,std::true_type{});
-    }
-    TableView( unsigned int iSize, std::array<void*, sizeof...(Args)>& iArray):
-    m_size(iSize),
-    m_values(iArray) {}
-
-    TableView( unsigned int iSize, std::array<void const*, sizeof...(Args)>& iArray):
-    m_size(iSize),
-    m_values(iArray) {}
-
-    unsigned int size() const {
-      return m_size;
-    }
-    
-    template<typename U>
-    typename U::type const& get(size_t iRow) const {
-      return static_cast<typename U::type*>(columnAddress<U>())+iRow;
-    }
-    
-    template<typename U>
-    ColumnValues<typename U::type> column() const {
-      return ColumnValues<typename U::type>{static_cast<typename U::type const*>(columnAddress<U>()), m_size};
-    }
-    
-    const_iterator begin() const { return const_iterator{m_values}; }
-    const_iterator end() const { return const_iterator{m_values,size()}; }
-    
-  private:
-    std::array<void const*, sizeof...(Args)> m_values;
-    unsigned int m_size;
-
-    template<typename U>
-    void const* columnAddress() const {
-      return m_values[impl::GetIndex<0,U,Layout>::index];
-    }
-
-    template <int I, typename T>
-    void fillArray( T const& iTable, std::true_type) {
-      using ElementType = typename std::tuple_element<I, Layout>::type;
-      m_values[I] = iTable.columnAddressWorkaround(static_cast<ElementType const*>(nullptr));
-      fillArray<I+1>(iTable, std::conditional_t<I+1<sizeof...(Args), std::true_type, std::false_type>{});
-    }
-    template <int I, typename T>
-    void fillArray( T const& iTable, std::false_type) {}
-
-    
-  };
-  
-  template<typename... T>
-  template<typename... U>
-  TableView<U...> Table<T...>::view() const {
-    return TableView<U...>{*this};
-  };
-  
   /* Table Type Manipulation */
   template <typename T1, typename T2> struct AddColumns;
   template <typename... T1, typename... T2>
